toolkit/components/passwordmgr/test/mochitest/test_autocomplete_basic_form_related_realms.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test login autocomplete with related realms</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
Login Manager test: related realms autocomplete

<script>
  addLoginsInParent(
    // Simple related domain relationship where example.com and other-example.com are in the related domains list
    ["https://other-example.com", "https://other-example.com", null, "relatedUser1", "relatedPass1", "uname", "pword"],

    // Example.com and example.co.uk are related, so sub.example.co.uk should appear on example.com's autocomplete dropdown
    // The intent is to cover the ebay.com/ebay.co.uk and all other country TLD cases
    // where the sign in page is actually signin.ebay.com/signin.ebay.co.uk but credentials could have manually been entered
    // for ebay.com/ebay.co.uk or automatically stored as signin.ebay.com/sigin.ebay.co.uk
    ["https://sub.example.co.uk", "https://sub.example.co.uk", null, "subUser1", "subPass1", "uname", "pword"],
    );
</script>
<p id="display"></p>
<div id="content">
  <form id="form1" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: related realms autocomplete. **/

const uname = getFormElementByName(1, "uname");
const pword = getFormElementByName(1, "pword");

function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function sendFakeAutocompleteEvent(element) {
  var acEvent = document.createEvent("HTMLEvents");
  acEvent.initEvent("DOMAutoComplete", true, false);
  element.dispatchEvent(acEvent);
}

function spinEventLoop() {
  return Promise.resolve();
}

async function promiseACPopupClosed() {
  return SimpleTest.promiseWaitForCondition(async () => {
    let popupState = await getPopupState();
    return !popupState.open;
  }, "Wait for AC popup to be closed");
}

add_setup(async () => {
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_form_related_domain_menuitems() {
  await SimpleTest.promiseFocus(window);
  restoreForm();
  const autocompleteItems = await openAutocompletePopup();
  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entires are selected upon opening");

  let expectedMenuItems = ["relatedUser1", "subUser1"];
  checkAutoCompleteResults(autocompleteItems, expectedMenuItems, window.location.host, "Check all menuitems are displayed correctly");

  let acEvents = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterPropsAC, clear: true });
  is(acEvents.length, 1, "One autocomplete event");
  checkACTelemetryEvent(acEvents[0], uname, {
    "hadPrevious": "0",
    "login": expectedMenuItems.length + "",
    "loginsFooter": "1"
  });
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by opening

  synthesizeKey("KEY_ArrowDown"); // first item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  is(pword.value, "relatedPass1", "password should match the login that was selected");
  checkLoginForm(uname, "relatedUser1", pword, "relatedPass1");

  restoreForm();

  await openAutocompletePopup();

  synthesizeKey("KEY_ArrowDown"); // first item
  synthesizeKey("KEY_ArrowDown"); // second item
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting

  synthesizeKey("KEY_Enter");
  await promiseFormsProcessedInSameProcess();
  is(pword.value, "subPass1", "password should match the login that was selected");
  checkLoginForm(uname, "subUser1", pword, "subPass1");

  restoreForm();
});

</script>
</pre>
</body>
</html>
